feat: ./Ex10.c; fixed: math errors in mean calc

+++ exercicios/Ex09.c
@@ -24,7 +24,7 @@ float media(int* array, int n){
         sum += array[i];
     }
     
-	media = sum/n;
+	media = (float)sum/n;
 	
 	return media;
 }
